requirements.txt
fastapi
uvicorn
Pillow
requests
pydantic
pydantic_settings
pydantic-core
streamlit
instructor
vllm-flash-attn